Deployment/requirements.txt
gradio
python-dotenv
langchain
langchain-openai
langchain-pinecone
openai
pinecone-client
